requirements.txt
chromadb==1.4.0
langchain==1.2.3
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-tavily==0.2.16
langchain-text-splitters==1.1.0
pypdf==6.5.0
python-dotenv==1.2.1
streamlit==1.52.2
